+++ src/RuleEngine.Tests/Utils/ExpressionExtensions/ConditionalExpressionTraceTests.cs
@@ -4,7 +4,7 @@ using RuleEngine.Utils;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace RuleEngine.Tests.Utils.ExressionExtensions
+namespace RuleEngine.Tests.Utils.ExpressionExtensions
 {
     public class ConditionalExpressionTraceTests
     {
