+++ src/RuleEngine.Tests/Utils/ExpressionExtensions/DynamicExpressionTraceTests.cs
@@ -8,7 +8,7 @@ using RuleEngine.Utils;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace RuleEngine.Tests.Utils.ExressionExtensions
+namespace RuleEngine.Tests.Utils.ExpressionExtensions
 {
     public class DynamicExpressionTraceTests
     {
