+++ src/RuleEngine.Tests/Utils/ExpressionExtensions/InvocationExpressionTraceTests.cs
@@ -5,7 +5,7 @@ using RuleEngine.Utils;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace RuleEngine.Tests.Utils.ExressionExtensions
+namespace RuleEngine.Tests.Utils.ExpressionExtensions
 {
     public class InvocationExpressionTraceTests
     {
