+++ src/RuleEngine.Tests/Utils/ExpressionExtensions/MethodCallExpressionTraceTests.cs
@@ -7,7 +7,7 @@ using SampleModel;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace RuleEngine.Tests.Utils.ExressionExtensions
+namespace RuleEngine.Tests.Utils.ExpressionExtensions
 {
     public class MethodCallExpressionTraceTests
     {
