+++ src/RuleEngine.Tests/Utils/ExpressionExtensions/UnaryExpressionTraceTests.cs
@@ -6,7 +6,7 @@ using RuleEngine.Utils;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace RuleEngine.Tests.Utils.ExressionExtensions
+namespace RuleEngine.Tests.Utils.ExpressionExtensions
 {
     public class UnaryExpressionTraceTests
     {
